feat  LZ rest API
Criação do RestController da LZ, junto com toda a estrutura dependente.(Repository,Dao,DTO,etc.).
Get pode ser acessado por localhost:8080/lz/{id}
Post pode ser acessado por localhost:8080/lz/load{body}

+++ donpipe/src/main/java/com/devxminds/donpipe/repositorios/EmpresaRepository.java
@@ -12,5 +12,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * @version 0.1
  *
  */
-public interface EmpresaRepository extends JpaRepository<Empresa, Integer> {
+public interface EmpresaRepository extends JpaRepository<Empresa, Long> {
 }
